+++ CHANGELOG.md
@@ -20,9 +20,10 @@ This changelog starts at version `0.20.0` and includes a selection of significan
   - 0.60.0
     - Breaking change of naming in `getFailedClients()` method: previously, it used the `includeBlankClients` parameter, which is now renamed to `includeBlank`.
     - Added following parameters to `getOnlineClients()` and `getOfflineClients()` method: `includeBlank`.
-    - Added following methods: `getUnseenClients()`.
+    - Added following methods: `getUnseenClients()`, `getStaleClients()`.
     - Matching empty clients now considers whether file backups or image backups are enabled.
     - Matching failed clients now considers whether file backups or image backups are enabled.
+    - Matching OK clients now considers whether file backups or image backups are enabled.
 
   - 0.54.0
     - Added following parameters to `getActivities()` and `getCurrentActivities()` method: `includePaused`.